Revert "Revert "UserPageLayout移動""
This reverts commit d316dc9dae1cba8c93f390b9ee2347ab364180b3.
