+++ src/app/users/[userKey]/layout.tsx
@@ -1,4 +1,6 @@
-import { UserPageLayout } from "@/components/layouts/UserPageLayout";
+import type { ReactNode } from "react";
+
+import { UserCard } from "./_components/UserCard";
 
 // TODO: https://github.com/vercel/next.js/issues/52126 が解決されたら追加
 
@@ -12,12 +14,16 @@ import { UserPageLayout } from "@/components/layouts/UserPageLayout";
 //   };
 // }
 
-export default function RootLayout({
-  children,
-  params,
-}: {
-  children: React.ReactNode;
-  params: { userKey: string };
-}) {
-  return <UserPageLayout userKey={params.userKey}>{children}</UserPageLayout>;
+type Props = {
+  userKey: string;
+  children: ReactNode;
+};
+
+export async function UserPageLayout({ userKey, children }: Props) {
+  return (
+    <>
+      <UserCard userKey={userKey} />
+      {children}
+    </>
+  );
 }
