--- src/components/layouts/UserPageLayout/index.tsx
@@ -1,17 +1,0 @@
-import type { ReactNode } from "react";
-
-import { UserCard } from "@/components/features/user/UserCard";
-
-type Props = {
-  userKey: string;
-  children: ReactNode;
-};
-
-export async function UserPageLayout({ userKey, children }: Props) {
-  return (
-    <>
-      <UserCard userKey={userKey} />
-      {children}
-    </>
-  );
-}
